# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowDeck productivity app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for the authenticated user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: goalId
          in: query
          required: false
          schema:
            type: integer
        - name: dueDate
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Toggle task completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found

  # ── HABITS ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, userId, title, priority, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        notes:
          type: ["string", "null"]
        goalId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [urgent, high, normal, low]
        status:
          type: string
          enum: [pending, completed, archived]
        dueDate:
          type: ["string", "null"]
        recurrence:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
        sortOrder:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        goalId:
          type: integer
        categoryId:
          type: integer
        priority:
          type: string
          enum: [urgent, high, normal, low]
        dueDate:
          type: string
        recurrence:
          type: string
        estimatedMinutes:
          type: integer
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        goalId:
          type: integer
        categoryId:
          type: integer
        priority:
          type: string
          enum: [urgent, high, normal, low]
        status:
          type: string
          enum: [pending, completed, archived]
        dueDate:
          type: string
        recurrence:
          type: string
        estimatedMinutes:
          type: integer
        sortOrder:
          type: integer

    # ── HABIT ─────────────────────────────────────────────────
...